requirements.txt
# ─────────────────────────────────────────────
#  Task Manager – Python dependencies
# ─────────────────────────────────────────────

# ── Web framework ─────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ── Settings ──────────────────────────────────
pydantic-settings==2.6.1

# ── Database ──────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0           # migrations for production
# For PostgreSQL (uncomment in production):
# psycopg2-binary==2.9.10

# ── Auth ──────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── HTTP client (Strike API) ──────────────────
httpx==0.28.0

# ── Razorpay SDK ──────────────────────────────
razorpay==1.4.2

# ── Testing ───────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0             # also used as async test client
